% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandLineInterface.R
\name{commandLineInterface}
\alias{commandLineInterface}
\title{Provide a command line interface to loopRuns and getSanityChecks}
\usage{
commandLineInterface(argv)
}
\arguments{
\item{argv}{A character vector containing the command line arguments.}
}
\value{
A string containing the status or sanity table.
}
\description{
This function is made to be called by the rs tool. It takes the command line arguments
supplied by the user when calling rs, evaluates them, and according to them prepares the 
arguments finally passed to loopRuns or getSanityChecks.
}
\author{
David Klein
}
