% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSanityChecks.R
\name{getSanityChecks}
\alias{getSanityChecks}
\title{getSanityChecks}
\usage{
getSanityChecks(dirs = NULL)
}
\arguments{
\item{dirs}{a vector of paths to REMIND runs.
When NULL, the latest AMTs are used (only works on PIK cluster)}
}
\description{
Get overview of sanity check results for a list of REMIND runs
}
\details{
Currently the following columns are printed:
\itemize{
\item SumErr Summation Errors found when running \code{remind2::convGDX2MIF}
\item RangeErr Range Errors found when running \code{remind2::convGDX2MIF}
\item FixingErr Fixing Errors found when running \code{piamInterfaces::fixOnRef}
\item MissingVar Missing Variables found when running \code{piamInterfaces::checkMissingVars} for ScenarioMIP
\item ProjSumErr Summation Errors found when running \code{piamInterfaces::checkSummations} for ScenarioMIP
\item ProjSumErrReg Regional Summation Errors found when running \code{piamInterfaces::checkSummationsRegional}
for ScenarioMIP
}
}
\author{
Falk Benke
}
