% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelstats-package.R
\docType{package}
\name{modelstats-package}
\alias{modelstats}
\alias{modelstats-package}
\title{modelstats: Run Analysis Tools}
\description{
A collection of tools to analyze model runs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/modelstats}
  \item Report bugs at \url{https://github.com/pik-piam/modelstats/issues}
}

}
\author{
\strong{Maintainer}: Anastasis Giannousakis \email{giannou@pik-potsdam.de}

Authors:
\itemize{
  \item Anastasis Giannousakis \email{giannou@pik-potsdam.de}
  \item Oliver Richters
  \item Simon Krogmann
}

}
