% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promptAndRun.R
\name{promptAndRun}
\alias{promptAndRun}
\title{promptAndRun}
\usage{
promptAndRun(mydir = ".", user = NULL, daysback = 3)
}
\arguments{
\item{mydir}{a dir or vector of dirs}

\item{user}{the user whose runs will be shown. For -t, the pattern used for matching}

\item{daysback}{integer defining the number of days -c will look back in time to find runs}
}
\description{
prompts runs, will be called by rs2
}
\examples{
\dontrun{
  promptAndRun()
}

}
\author{
Anastasis Giannousakis, Oliver Richters
}
