% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colRunType.R
\name{colRunType}
\alias{colRunType}
\title{colRunType}
\usage{
colRunType(mydir = ".")
}
\arguments{
\item{mydir}{Path to the folder(s) where the run(s) is(are) performed}
}
\description{
What is the type of this run?
}
\author{
Anastasis Giannousakis
}
