% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foundInSlurm.R
\name{foundInSlurm}
\alias{foundInSlurm}
\title{foundInSlurm}
\usage{
foundInSlurm(mydir = ".", user = NULL)
}
\arguments{
\item{mydir}{Path to the folder(s) where the run(s) is(are) performed}

\item{user}{the user whose runs will be sought for}
}
\description{
Is the run found in SLURM?
}
\author{
Anastasis Giannousakis, Oliver Richters
}
