% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunStatus.R
\name{getRunStatus}
\alias{getRunStatus}
\title{getRunStatus}
\usage{
getRunStatus(mydir = dir(), sort = "nf", user = NULL, detailed = TRUE)
}
\arguments{
\item{mydir}{Path to the folder(s) where the run(s) is(are) performed}

\item{sort}{how to sort (nf=newest first)}

\item{user}{the user whose runs will be shown}

\item{detailed}{(boolean, default: TRUE). If FALSE, avoids costly information gathering and returns less info}
}
\description{
Returns the current status of a run or a vector of runs
}
\examples{
\dontrun{

a <- getRunStatus(dir())
}

}
\author{
Anastasis Giannousakis
}
