% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSanityChecks.R
\name{getSanityChecks}
\alias{getSanityChecks}
\title{getSanityChecks}
\usage{
getSanityChecks(dirs)
}
\arguments{
\item{dirs}{a vector of paths to REMIND runs}
}
\description{
Get overview of sanity check results for a list of REMIND runs
}
\details{
Currently the following columns are printed
- SumErr Summation Errors found when running [remind2::convGDX2MIF]
- RangeErr Range Errors found when running [remind2::convGDX2MIF]
- FixingErr Fixing Errors found when running [piamInterfaces::fixOnRef]
- MissingVar Missing Variables found when running [piamInterfaces::checkMissingVars]
- ProjSumErr Summation Errors found when running [piamInterfaces::checkSummations] for Scenario MIP
- ProjSumErrReg Regional Summation Errors found when running [piamInterfaces::checkSummationsRegional] for Scenario MIP
}
\author{
Falk Benke
}
