% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltests.R
\name{modeltests}
\alias{modeltests}
\title{modeltests}
\usage{
modeltests(
  mydir = ".",
  gitdir = NULL,
  model = NULL,
  user = NULL,
  email = TRUE,
  compScen = TRUE,
  mattermostToken = NULL
)
}
\arguments{
\item{mydir}{Path to the folder where the model is found}

\item{gitdir}{path to the git clone that sends the report via email}

\item{model}{Model name}

\item{user}{the user that starts the job and commits the changes}

\item{email}{whether an email notification will be send or not}

\item{compScen}{whether compScen has to run or not}

\item{mattermostToken}{token used for mattermost notifications}
}
\description{
Runs a group of tests for a specific model. A "config/scenario_config.csv" file
(relative) to the main folder has to exist in the model, describing the
scenarios to be tested. Also, a gitlab repository is needed to push
the generated README.md for documentation and automated reporting
of test results.
}
\seealso{
\code{\link{package2readme}}
}
\author{
Anastasis Giannousakis, David Klein
}
