% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printOutput.R
\name{printOutput}
\alias{printOutput}
\title{printOutput}
\usage{
printOutput(string, len1stcol = 67)
}
\arguments{
\item{string}{String to print}

\item{len1stcol}{length of first column}
}
\description{
Returns the output for getRunStatus nicely
}
\author{
Anastasis Giannousakis
}
