% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printOutput.R
\name{printOutput}
\alias{printOutput}
\title{printOutput}
\usage{
printOutput(string, len1stcol = 67, lenCols = NULL, colSep = "   ")
}
\arguments{
\item{string}{String to print}

\item{len1stcol}{length of first column}

\item{lenCols}{vector with length information for all columns}

\item{colSep}{string separating the columns}
}
\description{
Returns the output for getRunStatus nicely
}
\author{
Anastasis Giannousakis
}
